# docs/openapi.yaml
openapi: 3.1.0
info:
  title: central-config admin API
  version: 0.1.0
  summary: The control plane's HTTP admin API — the only way configuration is written.
  description: |-
    This is the **control plane**. Admins and deploy pipelines write here; the
    database is the source of truth and every write is also written through to
    NATS JetStream KV. Consuming services never call any route in this document
    on the request path — they watch KV and serve from an in-memory cache.
    [`docs/CONSUMER_CONTRACT.md`](CONSUMER_CONTRACT.md) is the other half of the
    story, and `pkg/configclient` implements it for Go.

    The document describes the handlers as they behave, not as they might ideally
    be shaped. Several things here will look wrong until you know why:

    - **`enabled` is an integer on this API and a boolean in KV.** It is not a
      mistake and it is not being fixed — see the schema, and the README's
      [int/bool trap](../README.md#the-intbool-trap).
    - **The `PUT` routes carry the row id in the body, not in the path.** They
      are registered at collection paths (`/flags/values`, `/configs/values`,
      `/localization/values`) and read `id` out of the JSON.
    - **A single row fetched by id that lies outside the caller's token scope
      answers 404, not 403.** A 403 would confirm the row exists.
    - **Localization is asymmetric**: it creates at `/localization`, deletes at
      `/localization/{id}`, and updates at `/localization/values`.

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /environments:
    get:
      tags: [environments]
      operationId: listEnvironments
      summary: List environments
      description: |-
        Narrowed to the caller's scope by the environment's own id rather than by
        an `environmentId` field, because here the row *is* the environment.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: One page of environments.
          content:
            application/json:
              schema:
                type: [array, 'null']
                items:
                  $ref: '#/components/schemas/Environment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [environments]
      operationId: createEnvironment
      summary: Create an environment
      description: |-
        Full-scope tokens only. Adding an environment changes the set every other
        scope is expressed in, so a token narrowed to environments 1 and 2 cannot
        mint a third.

        Only `name` is read; it is trimmed, must be non-empty and at most 100
        characters — the column's width, and an empty name makes an unusable admin
        UI. Reference rows have no KV representation of their own, so nothing is
        published: keys appear when an appsettings, bundle or flag-value row does.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Environment'
            example:
              name: qa
      responses:
        '201':
          description: The stored row, read back by its natural key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /environments/{id}:
    delete:
      tags: [environments]
      operationId: deleteEnvironment
      summary: Delete an environment
      description: |-
        The one route where `{id}` is itself the environment the scope check runs
        against, rather than a row that happens to live in one.

        An environment still referenced by a flag value, an appsettings row or a
        bundle is refused with 409 rather than cascaded. Losing an environment by
        accident would take its whole configuration with it.
      parameters:
        - $ref: '#/components/parameters/RowId'
      responses:
        '204':
          $ref: '#/components/responses/Deleted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /microservices:
    get:
      tags: [microservices]
      operationId: listMicroservices
      summary: List microservices
      description: |-
        Not narrowed by scope: a microservice belongs to no environment, so there
        is nothing to narrow it by. Authentication still applies.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: One page of microservices.
          content:
            application/json:
              schema:
                type: [array, 'null']
                items:
                  $ref: '#/components/schemas/Microservice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [microservices]
      operationId: createMicroservice
      summary: Create a microservice
      description: |-
        Any valid token may issue this: the row belongs to no environment and
        publishes nothing to KV, so there is no scope for it to escape. Only
        `name` is read, on the same terms as an environment's.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Microservice'
            example:
              name: checkout-api
      responses:
        '201':
          description: The stored row, read back by its natural key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Microservice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /microservices/{id}:
    delete:
      tags: [microservices]
      operationId: deleteMicroservice
      summary: Delete a microservice
      description: |-
        Full-scope tokens only. The row itself belongs to no environment, but what
        points at it does — a scoped token deleting it would reach every
        environment at once — so the guard treats it as global.

        A microservice still referenced by an appsettings row or a bundle is
        refused with 409 rather than cascaded.
      parameters:
        - $ref: '#/components/parameters/RowId'
      responses:
        '204':
          $ref: '#/components/responses/Deleted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /configs/values:
    get:
      tags: [appsettings]
      operationId: listMicroserviceConfigs
      summary: List appsettings rows
      description: |-
        Both filters are optional and a zero id matches everything.

        A scoped token's page is filtered on the way out, after the database has
        already paged, so it can come back holding fewer than `limit` rows without
        that meaning the end of the collection. That is the cost of not teaching
        the domain handlers about credentials.
      parameters:
        - name: microserviceId
          in: query
          required: false
          description: Restrict to one microservice. Absent or `0` matches every one.
          schema:
            type: integer
            format: int64
            minimum: 0
          example: 1
        - name: environmentId
          in: query
          required: false
          description: Restrict to one environment. Absent or `0` matches every one.
          schema:
            type: integer
            format: int64
            minimum: 0
          example: 1
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: One page of appsettings rows.
          content:
            application/json:
              schema:
                type: [array, 'null']
                items:
                  $ref: '#/components/schemas/MicroserviceAppSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [appsettings]
      operationId: createMicroserviceConfig
      summary: Create an appsettings row
      description: |-
        `microserviceId`, `environmentId` and `settingsJson` are required; `id` and
        `updatedAt` are assigned by the database and anything sent in them is
        ignored.

        This is the create where the body *is* authoritative for the scope check:
        `environmentId` is what gets inserted, so a token is held to the
        environment it names.

        The tree is written to the database first and pushed to KV after. The KV
        write is best effort — a failure is logged, the request still succeeds, and
        the reconciler heals the drift within `RECONCILE_INTERVAL`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MicroserviceAppSettings'
            example:
              microserviceId: 2
              environmentId: 2
              settingsJson:
                http:
                  timeoutMs: 1500
      responses:
        '201':
          description: The stored row, read back by its natural key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroserviceAppSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags: [appsettings]
      operationId: updateMicroserviceConfig
      summary: Update an appsettings row — the id goes in the body
      description: |-
        **No `{id}` in the path.** The route is the collection and the row is named
        by `id` in the JSON body. `microserviceId`, `environmentId` and
        `settingsJson` are required too, and all three are rewritten: an update may
        move the row to a different (microservice, environment), which is a
        different KV key.

        When it does move, the key it used to feed is read *before* the write and
        deleted after — afterwards there is nothing left to derive it from, and
        leaving it behind means consumers in the old environment keep serving
        settings no row backs. Both environments are checked against the caller's
        scope, so a token cannot push a row out of an environment it holds into one
        it does not, or vice versa.

        The scope check reads the row's current environment from the database
        rather than trusting the body, because on a row-addressed route an
        `environmentId` in the body is a claim rather than a fact.

        Send `expectedUpdatedAt` to make the write conditional; see the schema.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MicroserviceAppSettings'
            example:
              id: 201
              microserviceId: 2
              environmentId: 1
              settingsJson:
                http:
                  timeoutMs: 2000
              expectedUpdatedAt: '2026-08-05T23:27:56.194028Z'
      responses:
        '200':
          description: The stored row as it is after the write.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroserviceAppSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /configs/values/{id}:
    get:
      tags: [appsettings]
      operationId: getMicroserviceConfig
      summary: Get one appsettings row
      description: |-
        `{id}` is the appsettings row id, not the microservice id — which is why a
        consumer hydrating over HTTP has to be told the row ids as deployment
        configuration, and why `GET /inventory` exists.
      parameters:
        - $ref: '#/components/parameters/RowId'
      responses:
        '200':
          description: The appsettings row.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroserviceAppSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFoundOrOutOfScope'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [appsettings]
      operationId: deleteMicroserviceConfig
      summary: Delete an appsettings row
      description: |-
        Removes the row and the KV key it fed. The key is purged here rather than
        left to the next full reconcile sweep, which would leave consumers serving
        deleted settings for up to a whole interval.
      parameters:
        - $ref: '#/components/parameters/RowId'
      responses:
        '204':
          $ref: '#/components/responses/Deleted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |-
        `Authorization: Bearer <secret>`, matched in constant time against every
        configured credential without an early exit, so neither the secret nor a
        token's position in the list leaks through response timing.

        Two variables configure it. `ADMIN_TOKENS=name:scope:secret,...` names
        credentials and scopes each to a `|`-separated list of environment ids, or
        `*` for all of them; `ADMIN_TOKEN` is the single shared full-scope form,
        recorded in the audit trail as the actor `shared`. **A secret may not
        contain a comma** — entries are split on commas before they are split on
        colons, so a comma inside a secret silently starts an entry, and possibly a
        full-scope token, of its own.

        A token's scope narrows reads as well as writes: listings return only rows
        in the environments it names, `/inventory` and `/audit` are narrowed the
        same way, and a single row fetched by id from outside the scope answers 404.

        With neither variable set **authentication is disabled entirely** and the
        service warns at startup. That is a development mode; every route then
        behaves as though it were called with a full-scope token.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: |-
        Page size. Absent gives 100. **A value of `0`, or anything above 500, is
        clamped to 500 rather than to the default** — a caller asking for too much
        still wants rows back. A negative or non-numeric value is a 400.

        **Where the page is taken differs by route**, and it decides whether a
        short page means the collection is exhausted:

        - `/flags/values`, `/configs/values`, `/localization` and `/environments`
          page in SQL and are then narrowed to the caller's scope on the way out,
          so a scoped token's page can hold fewer rows than this without the
          collection being exhausted. Page until you get back `[]`.
        - `/audit` is narrowed **in SQL**, in the same statement that pages it, so
          a short page there really is the end of the trail.
        - `/inventory` takes no page in SQL at all: the queries return every row
          and the handler pages in Go over the rows the token may see — so a
          scoped page is full whenever enough in-scope rows remain. It is also
          why that one route is rate limited.
        - `/microservices` and `/flags` page in SQL and are not narrowed at all;
          their rows belong to no environment.
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 100
      example: 100
    Offset:
      name: offset
      in: query
      required: false
      description: Rows to skip. Negative or non-numeric is a 400.
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
      example: 0
    RowId:
      name: id
      in: path
      required: true
      description: |-
        The row's primary key. Non-numeric answers 400 (`invalid id`) from the
        handler's own parse; zero or negative answers 400 from the service's
        validation.

        On a delete the write guard runs first and cannot resolve an unparseable
        id to an environment, so it fails closed and treats the request as
        reaching every one: a scoped token gets **403** where a full-scope token
        would have got the 400.

        This parameter is shared by seven routes over five tables, so the example
        below cannot be right for all of them — ids are per table. Against the
        seeded compose stack: environments and microservices are `1`–`3`, flags
        are `7`–`9`, flag values `100`–`103`, appsettings `200`–`202` and
        localization `300`–`301`. `GET /inventory` is the general answer.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
  responses:
    Deleted:
      description: |-
        Deleted. No body. Where the row fed a KV key that key is purged on the same
        request rather than left to the reconciler's sweep; the two reference
        tables have no key of their own to purge.
    BadRequest:
      description: |-
        The request could not be used as sent. `error` names what was wrong:
        a body that is not JSON or exceeds 1 MiB (`invalid request body`), a
        malformed page or filter parameter (`invalid query parameter`,
        `invalid environmentId`), an id that is zero or negative, a `settingsJson`
        or `bundleJson` that is not a JSON **object**, or a payload over the 512 KiB
        KV ceiling.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            notAnObject:
              summary: settingsJson is not an object
              description: |-
                A top-level array or scalar is well-formed JSON that would break
                every consumer's deserialization at once, so it is refused rather
                than published.
              value:
                error: invalid settings json
            tooLarge:
              summary: over the 512 KiB ceiling
              description: |-
                The same constant the KV buckets are provisioned with. Checked
                before the database write so a payload can never be accepted with a
                201 and then refused by JetStream forever afterwards.
              value:
                error: settings json exceeds the maximum value size
    Unauthorized:
      description: |-
        No bearer token, or one that matches no configured credential. Every route
        except `GET /health`, `GET /livez` and `GET /metrics` requires one — reads
        included, because a single unauthenticated GET was enough to walk off with
        every appsettings tree and bundle in the estate.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
    Forbidden:
      description: |-
        The token is valid but its scope does not cover what this write touches.
        A route marked full-scope-only (creating an environment, deleting a
        microservice or a flag definition) answers this for any scoped token, and so
        does a write whose target environment could not be determined at all — the
        guard fails closed.

        Note the asymmetry with reads: **writes** out of scope are 403, but a
        **single row read** out of scope is 404.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: token is not scoped to this environment
    NotFound:
      description: |-
        The addressed row does not exist, or a row it references does not. The
        parent checks run in the application rather than being left to the foreign
        keys, so "no such environment" arrives as this rather than as an opaque
        driver error behind a 500.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: environment not found
...
    Conflict:
      description: |-
        Two different situations share this code.

        A **natural key collision** — an environment or microservice name, a flag
        key, or the (microservice, environment) and (microservice, environment,
        locale) tuples — where the `error` names the collision. Deleting a reference
        row that something still points at lands here too: it is refused rather than
        cascaded, because an environment is not the kind of thing to remove by
        accident.

        A **lost optimistic-concurrency race**: the request carried
        `expectedUpdatedAt` and the stored row no longer holds that timestamp, so
        somebody else edited it first. Nothing was changed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            staleWrite:
              summary: somebody else edited it first
              value:
                error: flag value was modified by someone else
            alreadyExists:
              summary: the natural key is taken
              value:
                error: appsettings for this microservice and environment already exist
            stillReferenced:
              summary: the reference row is still pointed at
              value:
                error: environment still has flag, appsettings or localization rows
    RateLimited:
      description: |-
        The caller's budget is spent. `Retry-After` carries the whole seconds
        until the bucket refills.

        A write is charged twice: once at the edge by client address, before
        anything about the caller has been checked, and again per credential once
        the token is known. The edge charge goes to one of two separate
        namespaces depending on whether the request arrived with a credential
        this deployment configured. A **credential-less** write is charged four
        tokens against its own bucket — a quarter of the rate — so that a flood
        of them cannot spend the budget an authenticated caller at the same
        address is about to need, which behind an ingress means every operator at
        once.

        `WRITE_RATE_LIMIT_PER_MINUTE` (default 120) is the ceiling for each
        bucket, and a non-positive value disables the limiter entirely — it is a
        safety net, not something that should be able to lock an operator out of
        production by misconfiguration. Behind a proxy the address-keyed budget
        is the fleet's shared edge budget rather than one operator's, since the
        peer address is the same for everything arriving through it; the
        per-credential budget is what separates operators.

        **Reads are unmetered with one exception.** `GET /inventory` takes no
        page in SQL — it reads all three domains whole and pages in Go — so it
        draws on a budget of its own, keyed per credential in a namespace
        separate from the write budget, and can answer 429 like a write.
      headers:
        Retry-After:
          description: Whole seconds to wait before the next write will be accepted.
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 3
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: too many write requests
    ServerError:
      description: |-
        Something failed below the handler. The body is always this same opaque
        message: driver errors name tables, connection strings and file paths, so
        they go to the log instead.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: internal server error
  schemas:
    Error:
      type: object
      description: |-
        The uniform error envelope. Every non-2xx body a handler writes is this
        shape; the router's own 405 is the one exception, and it is described in
        the API description above rather than under an operation.
      properties:
        error:
          type: string
          description: A short human-readable reason, safe to show an operator.
      required: [error]
...
    Environment:
      type: object
      description: |-
        One environment. Its `id` is what token scopes, KV key prefixes and every
        other domain's `environmentId` refer to.
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: Assigned by the database.
        name:
          type: string
          maxLength: 100
          description: |-
            Trimmed, non-empty, at most 100 characters. Note the JSON name — the Go
            field and the column are both `ENVIRONMENT`.
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          description: Set to the database's `CURRENT_TIMESTAMP` on every write.
      example:
        id: 1
        name: dev
        updatedAt: '2026-08-05T23:27:56.194028Z'
    Microservice:
      type: object
      description: |-
        One microservice. It belongs to no environment — its appsettings and bundles
        do.
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: Assigned by the database.
        name:
          type: string
          maxLength: 100
          description: Trimmed, non-empty, at most 100 characters.
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      example:
        id: 1
        name: catalog-api
        updatedAt: '2026-08-05T23:27:56.194028Z'
...